Conditional loading of ctypes wintypes module
