+++ esphome/writer.py
@@ -13,7 +13,8 @@ from esphome.const import ARDUINO_VERSION_ESP32_1_0_0, ARDUINO_VERSION_ESP8266_2
     CONF_LOCAL, CONF_PLATFORMIO_OPTIONS, CONF_REPOSITORY, CONF_TAG, CONF_USE_CUSTOM_CODE
 from esphome.core import CORE, EsphomeError
 from esphome.core_config import GITHUB_ARCHIVE_ZIP, LIBRARY_URI_REPO, VERSION_REGEX
-from esphome.helpers import mkdir_p, run_system_command, symlink, islink, readlink, unlink
+from esphome.helpers import mkdir_p, run_system_command
+from esphome.symlink_ops import symlink, islink, readlink, unlink
 from esphome.pins import ESP8266_FLASH_SIZES, ESP8266_LD_SCRIPTS
 from esphome.py_compat import IS_PY3, string_types
 from esphome.storage_json import StorageJSON, storage_path
@@ -233,6 +234,7 @@ def symlink_esphome_core_version(esphome_core_version):
         # Remove symlink when changing back from local version
         if islink(dst_path):
             unlink(dst_path)
+
 
 def format_ini(data):
     content = u''
